Crate dkg_standalone_runtime

Re-exports

pub use constants::currency::*;

pub use constants::time::*;

Modules

api

constants

opaque

Opaque types. These are used by the CLI to instantiate machinery that don’t need to know the specifics of the runtime. They can then be made to be agnostic over specific formats of data like extrinsics, allowing for them to continue syncing the network through upgrades to even the core data structures.

Macros

construct_runtime

Construct a runtime, with the given name and the given pallets.

parameter_types

Create new implementations of the Get trait.

Structs

AdjustmentVariable

BalanceToU256

BetterUnsignedThreshold

BlockExecutionWeight

Time to execute an empty block. Calculated by multiplying the Average with 1 and adding 0.

BlockHashCount

BondingDuration

ChainIdentifier

CurrencyAdapter

Implements the transaction payment for a pallet implementing the Currency trait (eg. the pallet_balances) using an unbalance handler (implementing OnUnbalanced).

DKGAccountId

DecayPercentage

ElectionProviderBenchmarkConfig

The numbers configured here could always be more than the the maximum limits of staking pallet to ensure election snapshot will not run out of memory. For now, we set them to smaller values since the staking is bounded and the weight pipeline takes hours for this single pallet.

ExistentialDeposit

ExtrinsicBaseWeight

Time to execute a NO-OP extrinsic, for example System::remark. Calculated by multiplying the Average with 1 and adding 0.

GenesisConfig

IdentityFee

Implementor of WeightToFee that maps one unit of weight to one unit of fee.

IndexDeposit

MaxAuthorities

MaxElectingVoters

MaxLocks

MaxNominations

MaxNominatorRewardedPerValidator

MinPointsToBalance

MinerMaxLength

MinerMaxWeight

MinimumMultiplier

MinimumPeriod

MultiPhaseUnsignedPriority

NominationPoolsPalletId

NposSolution16

OffchainRepeat

OffendingValidatorsThreshold

Offset

OnChainSeqPhragmen

OperationalFeeMultiplier

Origin

The runtime origin type representing the origin of a call.

PalletId

A pallet identifier. These are per pallet and should be stored in a registry somewhere.

PalletInfo

Provides an implementation of PalletInfo to provide information about the pallet setup in the runtime.

Perbill

A fixed point representation of a number in the range [0, 1].

Period

Permill

A fixed point representation of a number in the range [0, 1].

PostUnbondPoolsWindow

ProposalLifetime

RefreshDelay

RewardCurve

RocksDbWeight

By default, Substrate uses RocksDB, so this will be the weight used throughout the runtime.

Runtime

RuntimeApi

RuntimeApiImpl

Implements all runtime apis for the client side.

RuntimeBlockLength

RuntimeBlockWeights

SS58Prefix

SessionsPerEra

SignedDepositBase

SignedDepositByte

SignedPhase

SignedRewardBase

SlashDeferDuration

StakingBenchmarkingConfig

StakingUnsignedPriority

StorageInfo

Metadata about storage from the runtime.

TargetBlockFullness

TargetedFeeAdjustment

A struct to update the weight multiplier per block. It implements Convert<Multiplier, Multiplier>, meaning that it can convert the previous multiplier to the next one. This should be called on on_finalize of a block, prior to potentially cleaning the weight data from the system pallet.

TimeToRestart

TransactionByteFee

U256ToBalance

UnsignedPhase

UnsignedProposalExpiry

Version

WebbMinerConfig

Enums

BalancesCall

Contains one variant per dispatchable that can be called by an extrinsic.

Call

DispatchClass

A generalized group of dispatch types.

Event

OriginCaller

StakerStatus

Indicates the initial status of the staker.

SystemCall

Contains one variant per dispatchable that can be called by an extrinsic.

TimestampCall

Contains one variant per dispatchable that can be called by an extrinsic.

Constants

MINER_MAX_ITERATIONS

Maximum number of iterations for balancing that will be executed in the embedded OCW miner of election provider multi phase.

VERSION

WASM_BINARY

WASM_BINARY_BLOATY

WEIGHT_PER_SECOND

Traits

BuildStorage

Complex storage builder stuff.

KeyOwnerProofSystem

Something which can compute and check proofs of a historical key owner and return full identification data of that key owner.

Randomness

A trait that is able to provide randomness.

StorageValue

A trait for working with macro-generated storage values under the substrate storage API.

Functions

native_version

The version information used to identify this runtime when compiled natively.

Type Definitions

AccountId

Some way of identifying an account on the chain. We intentionally make it equivalent to the public key of our transaction signing scheme.

AccountIndex

Address

The address format for describing accounts.

AllPalletsDeprecated

All pallets included in the runtime as a nested tuple of types.

AllPalletsReversedWithSystemFirst

All pallets included in the runtime as a nested tuple of types in reversed order. With the system pallet first.

AllPalletsWithSystem

All pallets included in the runtime as a nested tuple of types.

AllPalletsWithSystemReversed

All pallets included in the runtime as a nested tuple of types in reversed order.

AllPalletsWithoutSystem

All pallets included in the runtime as a nested tuple of types. Excludes the System pallet.

AllPalletsWithoutSystemReversed

All pallets included in the runtime as a nested tuple of types in reversed order. Excludes the System pallet.

Aura

AuraConfig

BagsList

Balances

BalancesConfig

Block

Block type as expected by this runtime.

CheckedExtrinsic

DKG

DKGConfig

DKGId

Identity of a DKG authority using ECDSA as its crypto.

DKGProposalHandler

DKGProposals

DKGProposalsConfig

ElectionProviderMultiPhase

Executive

Executive: handles dispatch to the various modules.

Grandpa

GrandpaConfig

Hash

A hash of some data used by the chain.

Header

Block header type as expected by this runtime.

Historical

Index

Index of a transaction in the chain.

Indices

IndicesConfig

Multiplier

Fee multiplier.

NominationPools

NominationPoolsConfig

RandomnessCollectiveFlip

Reputation

Reputation type

ResourceId

Resource ID type

Session

SessionConfig

Signature

Alias to 512-bit hash when used in the context of a transaction signature on the chain.

SignedExtra

The SignedExtension to the basic transaction logic.

SignedPayload

Staking

StakingConfig

Sudo

SudoConfig

System

SystemConfig

Timestamp

TransactionPayment

TransactionPaymentConfig

UncheckedExtrinsic

Unchecked extrinsic type as expected by this runtime.

Weight

Numeric range of a transaction weight.